Fix code constness in CBlockIndex::GetAncestor() overloads
Make the non-const overload of CBlockIndex::GetAncestor() reuse the
const overload implementation instead of the other way around. This way,
the constness of the const overload implementation is guaranteed. The
other way around, it was possible to implement the non-const overload in
a way which mutates the object, and since that implementation would be
called even for const objects (due to the reuse), we would get undefined
behavior.