[BUG] Check masternode/budget payments during block connection
As the tiertwo_reorg_mempool shows, the reorganization currently fails.
This is because, during a reorg, CheckBlock is called on blocks where
the prev is not connected to the chain yet.
Since the budget/mn payment is checked there, we call GeListForBlock
(requesting a list which has not been built yet) and so we erroneously
cache an empty list.

Fix this by moving IsBlockPayeeValid from CheckBlock to ConnectBlock
(as, to check budget/mn payments, the previous block deterministic list
must be fully built).

Also, to prevent sneaky bugs like this, in the future, only cache the
initial snapshot in GetListForBlock for the very last block before the
enforcement of DIP3.
After the enforcement block, at least the diff is always written to disk
in BuildListFromBlock.