[Tests] Proof of concept for P2CS vulnerability
The validation for P2CS scriptPubKey is incomplete and doesn't check all
the opcodes.
This opens a vulnerability.
A script could be crafted so that:
* it is identified as P2CS (passing IsPayToColdStaking), and recognized
by the wallet as ISMINE_SPENDABLE_DELEGATED.
* the assumed owner is not actually the owner (or not the only one) of
the coins.

In this Proof of concept, we craft a script that is recognized as own
P2CS by the owner wallet, but can actually be spent with **any** key.

This is achieved by including OP_DROP in a strategic position, so that,
during the script evaluation, part of the locking condition (included
only to fake IsPayToColdStaking) is removed, and replaced by a new
condition embedded in the spending scriptSig.