[GUI] Generate FAQ answer content programmatically
Each FAQ answer content is formatted as a self-contained HTML page for
styling purposes, however, having raw HTML in the `.ui` file's
`<string>` properties means that said HTML tags are also exported to
Transifex. This leads to a very ugly and confusing translation
experience that is prone to errors.

Here, we split up the content into shorter HTML-less strings: each list
item is it's own string, and each paragraph is it's own string. No HTML
tags are exported to Transifex.

The necessary HTML tags are added by way of helper functions that wrap
each string programmatically.