net_processing: Add missing erase orphan transactions when a block is connected.
Eliminating a leak that causes the orphan map to always grow to its maximum size.

Extra information:
I have skipped several PRs from upstream, this was first inside ConnectBlock, then was moved to ActivateBestChain and finally ended up moving it to net_processing under the SyncTransaction slot that we don't have anymore because we are much more updated.