BugFix: fix not updating GUI balance race condition.
The problem arises when we trigger the balance update worker task and the chain tip changes (due a reorg) before one of the threadpool's threads start executing the update task. As the tip in this situation has the same height that the one cached inside the model, the `processBalanceChangeInternal` returns true without setting the processing flag to false again. Blocking every future balance update.