[GUI] Ensure that all UI error/warning texts pass through translation
This adds translation where it was previously missing for some `UIError`
and `UIWarning` calls. in the case that the string is a pointer, the
base string has had translation added.

Also parameterized some translation strings to filter out string
literals that we DON'T want to be translated (eg '`-wallet`',
'`-paytxfee`', etc)