[Cleanup] Remove unused parameter in GetCurrentMasternode
Also pass the genesis hash instead of a null hash (as it was the
original behavior before bfa628e2db8faf8f20d55e71f9357a7353df4012) when
no-masternode-to-pay is found in GetLegacyMasternodeTxOut (e.g. when
there are less than 9 masternodes enabled on the network, and they are
all scheduled already).