[BUG] Fix broken tutorial wizard
The `-wallet` arguments can now represent paths to existing directories
or symlinks (by default it is empty, representing the path relative to
the datadir, i.e. the datadir itself).
It does not include the *filename* (unless explicitely specified in
`-wallet`) anymore.
Therefore the tutorial screen is never created (even on fresh installs).

Also, we are using `GetArg`, which returns the last argument from
command line (or first argument in conf file), and not checking all the
files with `GetArgs`.