Break addnode out from the outbound connection limits.
Adaptation of btc@50bd12ce0c49e574a5baf1a8df3a667810c6ad1e.

Previously addnodes were in competition with outbound connections
for access to the eight outbound slots.

One result of this is that frequently a node with several addnode
configured peers would end up connected to none of them, because
while the addnode loop was in its two minute sleep the automatic
connection logic would fill any free slots with random peers.
This is particularly unwelcome to users trying to maintain links
to specific nodes for fast block relay or purposes.

Another result is that a group of nine or more nodes which are
have addnode configured towards each other can become partitioned
from the public network.

This commit introduces a new limit of eight connections just for
addnode peers which is not subject to any of the other connection
limitations (including maxconnections).

The choice of eight is sufficient so that under no condition would
a user find themselves connected to fewer addnoded peers than
previously.  It is also low enough that users who are confused
about the significance of more connections and have gotten too
copy-and-paste happy will not consume more than twice the slot
usage of a typical user.

Any additional load on the network resulting from this will likely
be offset by a reduction in users applying even more wasteful
workaround for the prior behavior.

The retry delays are reduced to avoid nodes sitting around without
their added peers up, but are still sufficient to prevent overly
aggressive repeated connections.  The reduced delays also make
the system much more responsive to the addnode RPC.

Ban-disconnects are also exempted for peers added via addnode since
the outbound addnode logic ignores bans.  Previously it would ban
an addnode then immediately reconnect to it.

A minor change was also made to CSemaphoreGrant so that it is
possible to re-acquire via an object whos grant was moved.