[GUI] Force argument insertion when switching languages
On the Welcome screen, when users change the language, a call to Qt's
internal `retranslateUi()` is made. This, however does not take into
consideration any arguments in a source string, and instead treats them
as string literals (ie, `%1` is not considered an argument, but rather a
 literal string)

Fix this by adding a supplemental call to `setText()` that DOES allow
for argument processing directly after the retranslation.