TierTwo: do not ask for single MN items if the node is in initial synchronization state.
if it happens before or while the node's tier two initial synchronization is being executed, the peer will move to single items requests instead of requesting the full list.

Plus, decrease the DoS score for a bad mnb signature during the initial synchronization process, so we don't end up banning pre v5.3.3 nodes too fast for an invalidly serialized mnb (this issue was solved in #2611 removing the mnb BIP155 flag, this is only an extra guard to give a bit more time to non-updated peers to upgrade while the network is upgrading).