[GUI] Fix proposal tooltip menu location skewing
When there are enough proposals to need the GUI to scroll down to see
them all, the location of the tooltip menu gets skewed to the point of
it not being visible. This is because it's location was being based on
the initial position of the proposal card, even for off-screen ones.

This fixes the issue by taking the scrollarea's vertical scrollbar value
 into consideration.

Github-Pull: #2664
Rebased-From: 4ba2c4ca9567d912fbc63dd1650bc8e9560138d0